Update is-same-user-group.js
I think here it assumes that I am sending the `userGroup` in the payload, which I am not. Because of that, I am adding an additional check to see if the endpoint user is in fact trying to send over a `userGroup` or not. If not, then skip the code block.